+++ alikafka-20190916/src/Models/AddUserDefinedSgShrinkRequest.php
@@ -1,0 +1,76 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Alikafka\V20190916\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class AddUserDefinedSgShrinkRequest extends Model
+{
+    /**
+     * @var string
+     */
+    public $instanceId;
+
+    /**
+     * @var string
+     */
+    public $regionId;
+
+    /**
+     * @var string
+     */
+    public $sgIdListShrink;
+    protected $_name = [
+        'instanceId' => 'InstanceId',
+        'regionId' => 'RegionId',
+        'sgIdListShrink' => 'SgIdList',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->instanceId) {
+            $res['InstanceId'] = $this->instanceId;
+        }
+
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+
+        if (null !== $this->sgIdListShrink) {
+            $res['SgIdList'] = $this->sgIdListShrink;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['InstanceId'])) {
+            $model->instanceId = $map['InstanceId'];
+        }
+
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+
+        if (isset($map['SgIdList'])) {
+            $model->sgIdListShrink = $map['SgIdList'];
+        }
+
+        return $model;
+    }
+}