+++ alikafka-20190916/ChangeLog.md
@@ -1,3 +1,14 @@
+2026-04-17 Version: 2.14.0
+- Support API AddUserDefinedSg.
+- Support API DeleteUserDefinedSg.
+- Support API ModifyUserDefinedSg.
+- Update API CreatePrePayInstance: add request parameters ConfluentConfig.KsqlList.
+- Update API CreatePrePayOrder: add request parameters ConfluentConfig.KsqlList.
+- Update API GetInstanceList: add response parameters Body.InstanceList.$.ConfluentConfig.KsqlList.
+- Update API GetInstanceList: add response parameters Body.InstanceList.$.ConfluentInstanceComponents.$.internalId.
+- Update API UpgradePrePayOrder: add request parameters ConfluentConfig.KsqlList.
+
+
 2025-11-27 Version: 2.13.0
 - Support API DowngradePostPayOrder.
 - Support API DowngradePrePayOrder.